% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddwc.R
\name{loaddwc}
\alias{loaddwc}
\title{Load Data Warehouse from OnCore Cloud for demo purposes}
\usage{
loaddwc(data_warehouse)
}
\arguments{
\item{data_warehouse}{}
}
\value{
a dataframe
}
\description{
Load Data Warehouse from OnCore Cloud for demo purposes
}
