% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_mapping.R
\name{table_mapping}
\alias{table_mapping}
\title{Title: Table Mapping Function}
\usage{
table_mapping(table_name)
}
\arguments{
\item{table_name}{}
}
\value{
data_warehouse the base name of the data warehouse to pull from pins
}
\description{
Title: Table Mapping Function
}
